readjustment of thelibrary, cleaning unecessary
