+++ MazghanaGeometry.vcxproj
@@ -50,7 +50,9 @@
   <ItemGroup>
     <ClCompile Include="allTests.cpp" />
     <ClCompile Include="BSpline.cpp" />
+    <ClCompile Include="HalfeEdge.cpp" />
     <ClCompile Include="Maths.cpp" />
+    <ClCompile Include="MeshConvert.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
@@ -58,12 +60,14 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="point.cpp" />
+    <ClCompile Include="ReadObj.cpp" />
     <ClCompile Include="splineFunctions.cpp" />
     <ClCompile Include="TestHalfEdge.cpp" />
     <ClCompile Include="TestMaths.cpp" />
     <ClCompile Include="TestObjRead.cpp" />
     <ClCompile Include="testPoint.cpp" />
     <ClCompile Include="TestWriteOBJ.cpp" />
+    <ClCompile Include="WriteObj.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
